refactor(core): V1 extractions + partial V2 for #760
Extract modules to reduce file sizes toward ≤300 line goal:

V1 Quick Wins (6 files):
- _shared.py: extract StreamingState → _shared_streaming.py (257L)
- agent_store.py: extract BotAgentMapStore → bot_agent_map.py (275L)
- discord.py: restructure to discord/adapter.py + discord/lifecycle.py (281L)
- cli_streaming.py: extract CliStreamingParser → cli_streaming_parser.py (232L)
- turn_store.py: extract TurnStoreSessionMixin → infrastructure/turn_store_session.py (216L)
- command_router.py: reduce to 303L via consolidation

V2 Partial (3 extractions):
- outbound_router.py: extract TtsDispatch → outbound_tts.py (324L)
- hub_standalone.py: extract NATS wiring → nats_wiring.py (362L)
- cli_pool.py: extract CliPoolSessionMixin → cli_pool_session.py (406L)

Remaining work: hub.py (542L), full V2/V3/V4 extractions

All 2787 tests pass. Coverage: 82.96%
