+++ src/lyra/adapters/_shared.py
@@ -3,8 +3,9 @@
 Extracted from Telegram and Discord adapters to eliminate near-identical
 circuit-open / backpressure guard logic and reply_to_id parsing.
 
-Audio helpers live in _shared_audio; text utilities live in _shared_text.
-Both are re-exported here so existing importers continue to work without
+Audio helpers live in _shared_audio; text utilities live in _shared_text;
+streaming state classes live in _shared_streaming.
+All are re-exported here so existing importers continue to work without
 changes.
 """
 
@@ -13,7 +14,6 @@ from __future__ import annotations
 import asyncio
 import logging
 from collections.abc import Awaitable, Callable, Coroutine
-from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any
 
 from lyra.adapters._shared_audio import (
@@ -26,16 +26,17 @@ from lyra.adapters._shared_audio import (
     mime_to_ext,
 )
 
+# Re-exports from _shared_streaming — importers can use either module.
+from lyra.adapters._shared_streaming import (
+    STREAMING_EDIT_INTERVAL,
+    IntermediateTextState,
+    StreamState,
+)
+
 # Re-exports from _shared_audio — importers can use either module.
 from lyra.adapters._shared_text import chunk_text, sanitize_filename, truncate_caption
-from lyra.adapters.nats_stream_decoder import StreamChunkTimeout
 from lyra.core.circuit_breaker import CircuitRegistry
-from lyra.core.message import (
-    GENERIC_ERROR_REPLY,
-    InboundMessage,
-    Platform,
-)
-from lyra.core.render_events import TextRenderEvent
+from lyra.core.message import InboundMessage, Platform
 
 if TYPE_CHECKING:
     from lyra.adapters.outbound_listener import OutboundListener
@@ -152,10 +153,6 @@ ATTACHMENT_EXTS_BASE = frozenset(
 # Discord API message length limit — used by discord_formatting and discord_audio.
 DISCORD_MAX_LENGTH = 2000
 
-# Seconds between intermediate streaming edits (debounce).
-# Shared by Telegram and Discord adapters; aligned with each platform's rate limit.
-STREAMING_EDIT_INTERVAL = 1.0
-
 
 def resolve_msg(
     manager: "MessageManager | None", key: str, *, platform: str, fallback: str
@@ -203,178 +200,6 @@ class TypingTaskManager:
             task.cancel()
         if tasks:
             await asyncio.gather(*tasks, return_exceptions=True)
-
-
-# Maximum accumulated intermediate text length. Segments beyond this are
-# silently dropped — the display is a streaming placeholder, not a transcript.
-_MAX_INTERMEDIATE_CHARS = 8_000
-
-
-class IntermediateTextState:
-    """Tracks accumulated intermediate text and tool recap for streaming display.
-
-    Extracted from TelegramAdapter and DiscordAdapter to eliminate identical
-    intermediate-text accumulation and formatting logic. Each new segment gets
-    a ⏳ prefix and is separated by a newline so multiple thoughts stack
-    vertically. The latest tool recap can optionally be combined below the text.
-
-    Usage::
-
-        state = IntermediateTextState()
-        state.append(event.text)           # on TextRenderEvent(is_final=False)
-        state.set_tool_summary(recap)      # on ToolSummaryRenderEvent
-        display = state.display()          # text + recap (Telegram)
-        display = state.display(combine_recap=False)  # text only (Discord)
-    """
-
-    def __init__(self) -> None:
-        self._text: str = ""
-        self._tool_summary: str = ""
-
-    @property
-    def has_intermediate_text(self) -> bool:
-        """True once at least one intermediate segment has been appended."""
-        return bool(self._text)
-
-    @property
-    def text(self) -> str:
-        """Raw accumulated intermediate text (without tool summary)."""
-        return self._text
-
-    def append(self, new_text: str) -> None:
-        """Append a new intermediate segment with ⏳ prefix on a new line.
-
-        Silently drops segments once the accumulated text reaches
-        ``_MAX_INTERMEDIATE_CHARS`` — the placeholder is for live feedback,
-        not a full transcript.
-        """
-        if not new_text:
-            return
-        if len(self._text) >= _MAX_INTERMEDIATE_CHARS:
-            return
-        if self._text:
-            self._text += "\n⏳ " + new_text
-        else:
-            self._text = "⏳ " + new_text
-
-    def set_tool_summary(self, summary: str) -> None:
-        """Update the latest tool recap text."""
-        self._tool_summary = summary
-
-    def display(self, *, combine_recap: bool = True) -> str:
-        """Return the string to render in the placeholder message.
-
-        When *combine_recap* is ``True`` (default) and both intermediate text
-        and a tool recap are present, they are joined with a blank line so both
-        remain visible in the same message edit.  Pass ``combine_recap=False``
-        on Discord, where the recap lives in a separate embed.
-        """
-        if combine_recap and self._text and self._tool_summary:
-            return self._text + "\n\n" + self._tool_summary
-        return self._text or self._tool_summary
-
-
-_ERR_TIMEOUT_FALLBACK = (
-    "\u23f1\ufe0f The backend took longer than 120 s to respond. Please try again."
-)
-_ERR_NO_FINAL_FALLBACK = (
-    "\u26a0\ufe0f Response ended without a final message"
-    " (tool events only). Please try again."
-)
-
-
-def classify_stream_error(
-    stream_error: Exception | None,
-    *,
-    had_tool_events: bool,
-    final_text: str | None,
-    msg_fn: Callable[[str, str], str],
-) -> str | None:
-    """Return a descriptive error string for terminal error states.
-
-    Returns ``None`` when there is no error and a final text is present
-    (caller renders ``final_text`` normally).  Returns a user-facing string
-    for every error branch so callers never fall through to a bare
-    GENERIC_ERROR_REPLY silently.
-
-    Args:
-        stream_error:    Exception captured by the stream loop, or ``None``.
-        had_tool_events: Whether tool events were seen before the error.
-        final_text:      Final text captured from the stream, or ``None``.
-        msg_fn:          ``get_msg(key, fallback)`` callback for i18n.
-    """
-    if stream_error is not None:
-        if isinstance(stream_error, StreamChunkTimeout):
-            return msg_fn("error_timeout", _ERR_TIMEOUT_FALLBACK)
-        return msg_fn(
-            "error_stream",
-            f"\u26a0\ufe0f Streaming error:"
-            f" {type(stream_error).__name__}: {stream_error}. Please try again.",
-        )
-    if final_text is None and had_tool_events:
-        return msg_fn("error_no_final", _ERR_NO_FINAL_FALLBACK)
-    return None
-
-
-@dataclass
-class StreamState:
-    """Mutable event-loop state for send_streaming().
-
-    Extracted from Telegram and Discord adapters to eliminate the identical
-    7-variable state block, final-text capture, and display-text assembly.
-    Platform-specific rendering (API calls, text formatting) stays in each adapter.
-
-    Usage::
-
-        _st = StreamState()
-        async for event in events:
-            if isinstance(event, ToolSummaryRenderEvent):
-                _st.had_tool_events = True
-                # ... platform-specific tool render ...
-            else:
-                if event.is_final:
-                    _st.on_final_text(event)
-                else:
-                    _st.istate.append(event.text)
-                    # ... platform-specific intermediate edit ...
-        display_text = _st.build_display_text(adapter._msg)
-    """
-
-    had_tool_events: bool = False
-    istate: IntermediateTextState = field(default_factory=IntermediateTextState)
-    last_tool_edit: float | None = None
-    last_intermediate_edit: float | None = None
-    final_text: str | None = None
-    is_error_turn: bool = False
-    stream_error: Exception | None = None
-
-    def on_final_text(self, event: TextRenderEvent) -> None:
-        """Capture final text and error flag from a terminal TextRenderEvent."""
-        self.final_text = event.text
-        self.is_error_turn = event.is_error
-
-    def build_display_text(self, msg_fn: Callable[[str, str], str]) -> str | None:
-        """Assemble display text with error prefix and interrupt notice.
-
-        Returns ``None`` when no final text was received and no classified error
-        applies — callers then fall through to ``_deliver_final``'s error branch.
-        Returns a descriptive error string for timeout, stream error, or tool-only
-        turns so the user always sees a meaningful message.
-        """
-        if self.final_text is None:
-            return classify_stream_error(
-                self.stream_error,
-                had_tool_events=self.had_tool_events,
-                final_text=None,
-                msg_fn=msg_fn,
-            )
-        display = ("❌ " + self.final_text) if self.is_error_turn else self.final_text
-        if self.stream_error is not None:
-            if display:
-                display += msg_fn("stream_interrupted", " [response interrupted]")
-            else:
-                display = msg_fn("generic", GENERIC_ERROR_REPLY)
-        return display
 
 
 def parse_reply_to_id(reply_to_id: str | None) -> int | None:
