+++ src/lyra/core/cli_pool.py
@@ -17,6 +17,7 @@ if TYPE_CHECKING:
     from .stores.turn_store import TurnStore
 
 from .agent_config import ModelConfig
+from .cli_pool_session import CliPoolSessionMixin
 from .cli_pool_worker import (
     _LYRA_ROOT,
     CliPoolWorkerMixin,
@@ -42,7 +43,7 @@ __all__ = [
 log = logging.getLogger(__name__)
 
 
-class CliPool(CliPoolWorkerMixin):
+class CliPool(CliPoolSessionMixin, CliPoolWorkerMixin):
     """Pool of persistent Claude CLI processes (one per pool_id).
 
     Usage::
@@ -92,31 +93,6 @@ class CliPool(CliPoolWorkerMixin):
         # Updated by link_lyra_session() before each send, so the
         # _on_session_update callback can record {lyra_sid → cli_sid}.
         self._lyra_sessions: dict[str, str] = {}
-
-    def set_turn_store(self, store: TurnStore) -> None:
-        """Wire the TurnStore for CLI session persistence across restarts."""
-        self._turn_store = store
-
-    def link_lyra_session(self, pool_id: str, lyra_session_id: str) -> None:
-        """Associate the current Lyra session UUID with *pool_id*.
-
-        Called by the agent before each send so the persist callback can
-        map ``lyra_session_id → cli_session_id`` (for reply-to-resume).
-        """
-        self._lyra_sessions[pool_id] = lyra_session_id
-
-    def _persist_cli_session(self, pool_id: str, cli_session_id: str) -> None:
-        """Persist CLI session ID to TurnStore for --resume after daemon restart."""
-        if not cli_session_id or not _SESSION_ID_RE.match(cli_session_id):
-            return
-        lyra_sid = self._lyra_sessions.get(pool_id)
-        if lyra_sid and self._turn_store is not None:
-            try:
-                asyncio.get_running_loop().create_task(
-                    self._turn_store.set_cli_session(lyra_sid, cli_session_id)
-                )
-            except RuntimeError:
-                pass  # no running loop — test context without TurnStore
 
     async def start(self) -> None:
         """Start the idle reaper background task."""
