+++ src/lyra/infrastructure/stores/__init__.py
@@ -2,6 +2,7 @@
 
 from lyra.infrastructure.stores.agent_store import AgentStore
 from lyra.infrastructure.stores.auth_store import AuthStore
+from lyra.infrastructure.stores.bot_agent_map import BotAgentMapStore
 from lyra.infrastructure.stores.credential_store import (
     BotSecretRow,
     CredentialStore,
@@ -13,6 +14,7 @@ from lyra.infrastructure.stores.sqlite_base import SqliteStore
 __all__ = [
     "AgentStore",
     "AuthStore",
+    "BotAgentMapStore",
     "BotSecretRow",
     "CredentialStore",
     "IdentityAliasStore",
