+++ tools/file_exemptions.txt
@@ -2,13 +2,13 @@
 # Each entry must have a tracking issue.
 # Format: <path>  # <lines> lines — <issue> <description>
 
-src/lyra/core/hub/hub.py                      # 791 lines — #396 refactor backlog
-src/lyra/core/hub/outbound_router.py          # 387 lines — #753 dispatch extraction
+src/lyra/core/hub/hub.py                      # 542 lines — #396 refactor backlog
+src/lyra/core/hub/outbound_router.py          # 397 lines — #753 dispatch extraction
 src/lyra/bootstrap/hub_standalone.py          # 432 lines — #396 refactor backlog
 src/lyra/adapters/_shared.py                  # 432 lines — #396 refactor backlog
 src/lyra/core/cli_pool.py                     # 430 lines — #396 refactor backlog
-src/lyra/infrastructure/stores/agent_store.py # 341 lines — #396 refactor backlog
+src/lyra/infrastructure/stores/agent_store.py # 340 lines — #396 refactor backlog
 src/lyra/adapters/discord.py                  # 322 lines — #196 adapter protocol
-src/lyra/core/commands/command_router.py      # 310 lines — #396 refactor backlog
-src/lyra/core/cli_streaming.py                # 304 lines — #753 new extraction
+src/lyra/core/commands/command_router.py      # 316 lines — #396 refactor backlog
+src/lyra/core/cli_streaming.py                # 305 lines — #753 new extraction
 src/lyra/core/stores/turn_store.py            # 301 lines — pending migration
